+++ tests/epyccel/test_arrays.py
@@ -27,7 +27,7 @@ def check_array_equal(a, b):
 #==============================================================================
 
 def test_array_assigned_dtype(language):
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max,   dtype=int)
+    integer   = randint(low = iinfo('int32').min,   high = iinfo('int32').max)
     integer8  = randint(low = iinfo('int8').min,  high = iinfo('int8').max,  dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max, dtype=np.int32)
@@ -2267,7 +2267,7 @@ def test_argument_negative_index_2(language):
 
 
 def test_c_order_argument_negative_index(language):
-    a = np.random.randint(20, size=(3,4))
+    a = np.array(np.random.randint(20, size=(3,4)), dtype=int)
 
     f1 = arrays.test_c_order_argument_negative_index
     f2 = epyccel(f1, language = language)
@@ -2275,7 +2275,7 @@ def test_c_order_argument_negative_index(language):
 
 
 def test_f_order_argument_negative_index(language):
-    a = np.array(np.random.randint(20, size=(3,4)), order='F')
+    a = np.array(np.random.randint(20, size=(3,4)), order='F', dtype=int)
 
     f1 = arrays.test_f_order_argument_negative_index
     f2 = epyccel(f1, language = language)
@@ -4258,7 +4258,7 @@ def test_array_ndmin_1(language):
     a = arrays.a_1d
     b = arrays.a_2d_c
     c = arrays.a_2d_c
-    d = randint(low = iinfo(int).min, high = iinfo(int).max, dtype=int, size=(2,3,4))
+    d = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max, size=(2,3,4)), dtype=int)
     e = d.copy(order='F')
 
     check_array_equal(f1(a), f2(a))
@@ -4287,7 +4287,7 @@ def test_array_ndmin_2(language):
     a = arrays.a_1d
     b = arrays.a_2d_c
     c = arrays.a_2d_c
-    d = randint(low = iinfo(int).min, high = iinfo(int).max, dtype=int, size=(2,3,4))
+    d = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max, size=(2,3,4)), dtype=int)
     e = d.copy(order='F')
 
     check_array_equal(f1(a), f2(a))
@@ -4316,7 +4316,7 @@ def test_array_ndmin_4(language):
     a = arrays.a_1d
     b = arrays.a_2d_c
     c = arrays.a_2d_c
-    d = randint(low = iinfo(int).min, high = iinfo(int).max, dtype=int, size=(2,3,4))
+    d = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max, size=(2,3,4)), dtype=int)
     e = d.copy(order='F')
 
     check_array_equal(f1(a), f2(a))
@@ -4345,7 +4345,7 @@ def test_array_ndmin_2_order(language):
     a = arrays.a_1d
     b = arrays.a_2d_c
     c = arrays.a_2d_c
-    d = randint(low = iinfo(int).min, high = iinfo(int).max, dtype=int, size=(2,3,4))
+    d = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max, size=(2,3,4)), dtype=int)
     e = d.copy(order='F')
 
     check_array_equal(f1(a), f2(a))
@@ -4363,8 +4363,8 @@ def test_dtype_conversion_to_bool_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4401,7 +4401,7 @@ def test_dtype_conversion_to_int8_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
     integer8   = randint(low = iinfo('int8').min,   high = iinfo('int8').max, size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
@@ -4439,8 +4439,8 @@ def test_dtype_conversion_to_int16_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4477,8 +4477,8 @@ def test_dtype_conversion_to_int32_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4515,8 +4515,8 @@ def test_dtype_conversion_to_int64_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4552,8 +4552,8 @@ def test_dtype_conversion_to_float32_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4588,8 +4588,8 @@ def test_dtype_conversion_to_float64_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4623,8 +4623,8 @@ def test_dtype_conversion_to_complex64_from_other_types(language):
     size = (2, 2)
 
     bl = randint(0, 2, size = size, dtype= bool)
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4658,8 +4658,8 @@ def test_dtype_conversion_to_complex128_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4696,7 +4696,7 @@ def test_dtype_conversion_to_pyint_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
     integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
@@ -4733,8 +4733,8 @@ def test_dtype_conversion_to_pyfloat_from_other_types(language):
 
     bl = randint(0, 2, size = size, dtype= bool)
 
-    integer   = randint(low = iinfo('int').min,   high = iinfo('int').max, size = size, dtype=int)
-    integer8 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
+    integer   = np.array(randint(low = iinfo('int32').min,   high = iinfo('int32').max, size = size), dtype=int)
+    integer8  = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer16 = randint(low = iinfo('int16').min, high = iinfo('int16').max , size = size, dtype=np.int16)
     integer32 = randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size, dtype=np.int32)
     integer64 = randint(low = iinfo('int64').min, high = iinfo('int64').max , size = size, dtype=np.int64)
@@ -4769,9 +4769,9 @@ def test_dtype_conversion_to_pyfloat_from_other_types(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_bool(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -4842,9 +4842,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_bool(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int8(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -4915,9 +4915,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int8(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int16(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -4987,9 +4987,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int16(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int32(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5060,9 +5060,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int32(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int64(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5132,9 +5132,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int64(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_float32(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5203,9 +5203,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_float32(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_float64(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5274,9 +5274,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_float64(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_cfloat(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5344,9 +5344,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_cfloat(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_cdouble(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5417,9 +5417,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_cdouble(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_pyint(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5490,9 +5490,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_pyint(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_pyfloat(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5563,9 +5563,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_pyfloat(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_bool_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5636,9 +5636,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_bool_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int8_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5709,9 +5709,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int8_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int16_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5782,9 +5782,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int16_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int32_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5855,9 +5855,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int32_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_int64_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5927,9 +5927,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_int64_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_float32_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -5998,9 +5998,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_float32_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_float64_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -6069,9 +6069,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_float64_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_cfloat_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -6139,9 +6139,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_cfloat_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_cdouble_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -6212,9 +6212,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_cdouble_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_pyint_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
@@ -6285,9 +6285,9 @@ def test_src_dest_array_diff_sizes_dtype_conversion_to_pyint_orderF(language):
 def test_src_dest_array_diff_sizes_dtype_conversion_to_pyfloat_orderF(language):
     size = (1,2)
 
-    integer_1 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_2 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
-    integer_3 = randint(low = iinfo('int').min, high = iinfo('int').max , size = size, dtype=int)
+    integer_1 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_2 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
+    integer_3 = np.array(randint(low = iinfo('int32').min, high = iinfo('int32').max , size = size), dtype=int)
 
     integer8_1 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)
     integer8_2 = randint(low = iinfo('int8').min, high = iinfo('int8').max , size = size, dtype=np.int8)