+++ tests/epyccel/test_builtins.py
@@ -10,12 +10,12 @@ from pyccel.decorators import template
 ATOL = 1e-15
 RTOL = 2e-14
 
+# Use int32 for Windows compatibility
+min_int = iinfo(np.int32).min
+max_int = iinfo(np.int32).max
 
-min_int = iinfo('int').min
-max_int = iinfo('int').max
-
-min_float = finfo('float').min
-max_float = finfo('float').max
+min_float = finfo(float).min
+max_float = finfo(float).max
 
 def test_abs_i(language):
     def f1(x : 'int'):